+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Removed `experimental` from coordination API
+* Added `WithReaderLogContext`, `WithWriterLogContext` options to topic reader/writer to supply log entries with user context fields
 
 ## v3.108.3
 * Fixed handling of zero values for DyNumber